+++ disco/src/main/java/com/github/lawben/disco/aggregation/functions/MaxAggregateFunction.java
@@ -2,21 +2,21 @@ package com.github.lawben.disco.aggregation.functions;
 
 import com.github.lawben.disco.aggregation.DistributiveAggregateFunction;
 
-public class MaxAggregateFunction implements DistributiveAggregateFunction<Integer> {
+public class MaxAggregateFunction implements DistributiveAggregateFunction<Long> {
     @Override
-    public Integer lift(Integer inputTuple) {
+    public Long lift(Long inputTuple) {
         return inputTuple;
     }
 
     @Override
-    public Integer combine(Integer partialAggregate1, Integer partialAggregate2) {
+    public Long combine(Long partialAggregate1, Long partialAggregate2) {
         if (partialAggregate1 == null) return partialAggregate2;
         if (partialAggregate2 == null) return partialAggregate1;
         return Math.max(partialAggregate1, partialAggregate2);
     }
 
     @Override
-    public Integer lower(Integer aggregate) {
+    public Long lower(Long aggregate) {
         return aggregate;
     }
 }
